Separate file actions into own mixins
