+++ apps/files/src/default.js
@@ -18,7 +18,6 @@ const store = require('./store')
 function $gettext(msg) {
   return msg
 }
-const filesConfig = window.phoenixConfig.files || []
 
 const appInfo = {
   name: $gettext('Files'),
@@ -26,7 +25,6 @@ const appInfo = {
   icon: 'folder',
   isFileEditor: false,
   extensions: [],
-  fileActions: filesConfig.actions || [],
   fileSideBars: [
     {
       app: 'files-version',
